moved all highlight commands in same class
