+++ src/main/java/de/slimecloud/slimeball/main/SlimeBot.java
@@ -32,7 +32,7 @@ import de.slimecloud.slimeball.features.github.GitHubAPI;
 import de.slimecloud.slimeball.features.highlights.Highlight;
 import de.slimecloud.slimeball.features.highlights.HighlightListener;
 import de.slimecloud.slimeball.features.highlights.HighlightTable;
-import de.slimecloud.slimeball.features.highlights.commands.HighlightCommand;
+import de.slimecloud.slimeball.features.highlights.HighlightCommand;
 import de.slimecloud.slimeball.features.level.Level;
 import de.slimecloud.slimeball.features.level.LevelTable;
 import de.slimecloud.slimeball.features.level.LevelUpListener;
